refactor(cli): split cmdcore/register.go into flow/setup/claim/derive (E7)
Break the 34KB register.go into focused files: register.go keeps the
command + flag/prompt gating (registerE); register_flow.go holds the
DCR + approval-wait state machine (registerV2/waitForApprovalV2 + poll
cadence, saveRegState, next-steps output); register_setup.go holds the
fresh-machine trio (RegisterV2Setup/RegisterV2Active/SetupValues/prompt);
register_claim.go holds the claim affordance + console/claim URLs; and
register_derive.go holds the URL/name derivation helpers. Intra-package
moves only.
